[Syntax] Support Twig expressions in CSS.
This commit introduces a new syntax called `CSS (Twig)` which just extends
CSS and add Twig contexts. We then hook into specific contexts in HTML
to allow this syntax to take over instead of the default CSS syntax. This
will ensure that Twig expressions are not marked as illegal within CSS
blocks.

Resolves #25