Split snippets into block- and tag-snippets
Adds snippets for individual tags so typing a tag name and hitting tab key
expands to an individual tag enclosed in `{% ... %}`:

   if| =>  {% if | %}
   embed| =>  {% embed | %}

Triggers of existing block snippets are prefixed with `b` to avoid ambiguities
with tag-snippets. Otherwise it would be unclear whether blocks or tags are
expanded and/or completion panel was displayed requiring arrow key navigation.

   bif|

   {% if conditon %}
      |
   {% endif %}

Reorganize snippets in individual sub-directories in case someone wants to
disable individual ones based on type.